Object oriented design and inheritance basic ..
