+++ src/backend/components/py.pi/caliopen_pi/qualifiers/twitter.py
@@ -3,80 +3,27 @@
 from __future__ import absolute_import, print_function, unicode_literals
 import logging
 
-from caliopen_main.message.parameters import (NewInboundMessage,
-                                              Participant,
-                                              Attachment)
+from caliopen_main.message.parameters import NewInboundMessage
 from caliopen_main.message.parsers.twitter import TwitterDM
 from caliopen_main.discussion.core import Discussion
 from caliopen_main.discussion.core import (DiscussionThreadLookup,
                                            DiscussionListLookup,
                                            DiscussionGlobalLookup)
-from caliopen_storage.exception import NotFound
 
 from ..features import marshal_features
+from .base import BaseQualifier
 
 log = logging.getLogger(__name__)
 
 
-class UserDMQualifier(object):
-    """
-    Process a Twitter direct message to unmarshal it in our stack
-    """
+class UserDMQualifier(BaseQualifier):
+    """Process a Twitter direct message to unmarshal it in our stack."""
 
     _lookups = {
         'global': DiscussionGlobalLookup,
         'thread': DiscussionThreadLookup,
         'list': DiscussionListLookup,
     }
-
-    def __init__(self, user, identity):
-        self.user = user
-        self.identity = identity
-
-    def _get_tags(self, message):
-        """Evaluate user rules to get all tags for a mail."""
-        tags = []
-        if message.privacy_features.get('is_internal', False):
-            # XXX do not hardcode the wanted tag
-            internal_tag = [x for x in self.user.tags if x.name == 'internal']
-            if internal_tag:
-                tags.append(internal_tag[0].name)
-        message.tags = tags
-
-    def lookup(self, sequence):
-        """Process lookup sequence to find discussion to associate."""
-        log.debug('Lookup sequence %r' % sequence)
-        for prop in sequence:
-            try:
-                kls = self._lookups[prop[0]]
-                log.debug('Will lookup %s with value %s' %
-                          (prop[0], prop[1]))
-                return kls.get(self.user, prop[1])
-            except NotFound:
-                log.debug('Lookup type %s with value %s failed' %
-                          (prop[0], prop[1]))
-
-        return None
-
-    def create_lookups(self, sequence, message):
-        """Initialize lookup classes for the related sequence."""
-        for prop in sequence:
-            kls = self._lookups[prop[0]]
-            params = {
-                kls._pkey_name: prop[1],
-                'discussion_id': message.discussion_id
-            }
-            lookup = kls.create(self.user, **params)
-            log.info('Create lookup %r' % lookup)
-
-    def get_participant(self, participant):
-        """TODO: find related contact"""
-        p = Participant()
-        p.address = participant.address
-        p.type = participant.type
-        p.label = participant.label
-        p.protocol = "twitter"
-        return p
 
     def lookup_discussion_sequence(self, message, *args, **kwargs):
         """Return list of lookup type, value from a tweet."""
@@ -90,7 +37,8 @@ class UserDMQualifier(object):
         return seq
 
     def process_inbound(self, raw):
-        """Process inbound message.
+        """
+        Process inbound message.
 
         @param raw: a RawMessage object
             which should be a json conforming to
@@ -98,7 +46,6 @@ class UserDMQualifier(object):
             sending-and-receiving/guides/message-create-object
         @rtype: NewMessage
         """
-
         message = TwitterDM(raw.raw_data)
         new_message = NewInboundMessage()
         new_message.raw_msg_id = raw.raw_msg_id
@@ -114,7 +61,7 @@ class UserDMQualifier(object):
 
         participants = []
         for p in message.participants:
-            participant = self.get_participant(p)
+            participant, contact = self.get_participant(p)
             new_message.participants.append(participant)
             participants.append(participant)
         if not participants:
